Move Material3D conversion editor plugins to their own folder
